server/requirements.txt
# RadioWall Dependencies
# Install with: pip install -r requirements.txt

# Core
requests>=2.28.0          # HTTP client for Radio.garden API
geopy>=2.3.0              # Geographic distance calculations

# Touch Input (Linux)
evdev>=1.6.0              # USB HID event reading on Linux

# MQTT
paho-mqtt>=1.6.0          # MQTT client for touch event communication

# UPnP/DLNA Streaming
async-upnp-client>=0.38.0 # UPnP device discovery and control

# Configuration
pyyaml>=6.0               # YAML config file parsing

# Optional: Development/Testing
pytest>=7.0.0             # Testing framework
python-dotenv>=1.0.0      # Environment variable management

# Note: The old requirements included many Jupyter/matplotlib dependencies
# for the prototype. Those are no longer needed for the headless service.
# If you want to run the old playground.ipynb, install additionally:
#   pip install jupyter matplotlib pandas numpy python-vlc pynput
